Refactor out MTColorSchemeManager and replace with static invocations
