+++ src/com/chrisrm/idea/MTColorScheme.java
@@ -1,13 +1,7 @@
 package com.chrisrm.idea;
 
-import com.chrisrm.idea.actions.MTSwitchTheme;
 import com.intellij.openapi.components.ApplicationComponent;
-import com.intellij.openapi.editor.colors.EditorColorsManager;
 import org.jetbrains.annotations.NotNull;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 
 public class MTColorScheme implements ApplicationComponent {
 
@@ -16,7 +10,7 @@ public class MTColorScheme implements ApplicationComponent {
 
     @Override
     public void initComponent() {
-        new MTSwitchTheme().setTheme(new MTDataLayer().getValue("theme", "default"));
+        MTThemeUtil.setTheme(new MTDataLayer().getValue("theme", "default"));
     }
 
     public void disposeComponent() {
