+++ src/com/chrisrm/idea/actions/MTDarkerTheme.java
@@ -1,5 +1,6 @@
 package com.chrisrm.idea.actions;
 
+import com.chrisrm.idea.MTThemeUtil;
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 
@@ -8,11 +9,11 @@ import javax.swing.*;
 
 public class MTDarkerTheme extends AnAction {
 
-    String name = "Darker";
+    private static final String NAME = "Darker";
 
     @Override
     public void actionPerformed(AnActionEvent anActionEvent) {
         JOptionPane.showMessageDialog(null, "Keep in mind, the darker theme is a work in progress and there will be some issues.", "Work in progress", JOptionPane.INFORMATION_MESSAGE);
-        new MTSwitchTheme().setTheme(this.name);
+        MTThemeUtil.setTheme(NAME);
     }
 }
