+++ src/com/chrisrm/idea/actions/MTDefaultTheme.java
@@ -1,15 +1,16 @@
 package com.chrisrm.idea.actions;
 
+import com.chrisrm.idea.MTThemeUtil;
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 
 
 public class MTDefaultTheme extends AnAction {
 
-    String name = "Default";
+    private static final String NAME = "Default";
 
     @Override
     public void actionPerformed(AnActionEvent anActionEvent) {
-        new MTSwitchTheme().setTheme(this.name);
+        MTThemeUtil.setTheme(NAME);
     }
 }
