+++ src/com/chrisrm/idea/actions/MTLighterTheme.java
@@ -1,5 +1,6 @@
 package com.chrisrm.idea.actions;
 
+import com.chrisrm.idea.MTThemeUtil;
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 
@@ -8,11 +9,11 @@ import javax.swing.*;
 
 public class MTLighterTheme extends AnAction {
 
-    String name = "Lighter";
+    private static final String NAME = "Lighter";
 
     @Override
     public void actionPerformed(AnActionEvent anActionEvent) {
         JOptionPane.showMessageDialog(null, "Keep in mind, the lighter theme is a work in progress and there will be some issues.", "Work in progress", JOptionPane.INFORMATION_MESSAGE);
-        new MTSwitchTheme().setTheme(this.name);
+        MTThemeUtil.setTheme(NAME);
     }
 }
